requirements.txt
itchat==1.3.10
selenium==3.141.0
Appium_Python_Client==1.0.1
